+++ hahomematic/client/__init__.py
@@ -20,6 +20,7 @@ from hahomematic.const import (
     EVENT_SECONDS_SINCE_LAST_EVENT,
     HOMEGEAR_SERIAL,
     INIT_DATETIME,
+    INTERFACES_SUPPORTING_FIRMWARE_UPDATES,
     VIRTUAL_REMOTE_TYPES,
     Backend,
     CallSource,
@@ -136,6 +137,11 @@ class Client(ABC):
     @abstractmethod
     def supports_ping_pong(self) -> bool:
         """Return the supports_ping_pong info of the backend."""
+
+    @property
+    def supports_firmware_updates(self) -> bool:
+        """Return the supports_ping_pong info of the backend."""
+        return self.interface in INTERFACES_SUPPORTING_FIRMWARE_UPDATES
 
     async def proxy_init(self) -> ProxyInitState:
         """Init the proxy has to tell the CCU / Homegear where to send the events."""