+++ changelog.md
@@ -3,6 +3,7 @@
 - Add getLinkPeers XmlRPC method
 - Do not create update entities that are not updatable (manually remove obsolete update entities)
 - Only try device update refresh if device is updatable
+- Refactor update entity
 
 # Version 2024.8.15 (2024-08-29)